src/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
Jinja2==3.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
azure-identity==1.15.0 